requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
defusedxml==0.8.0rc2
Django==5.1.1
django-autoslug==1.9.9
django-countries==7.6.1
django-debug-toolbar==4.4.6
django-phonenumber-field==8.0.0
frozenlist==1.5.0
geoip2==4.8.0
idna==3.10
MarkupSafe==2.1.5
maxminddb==2.6.2
multidict==6.1.0
oauthlib==3.2.2
phonenumbers==8.13.50
pillow==10.4.0
progressbar2==4.5.0
propcache==0.2.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
python-utils==3.9.0
python3-openid==3.2.0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==75.5.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
Unidecode==1.3.8
urllib3==2.2.3
yarl==1.17.1
